[StableHLO][NFC] Factor out dot prod lowering patterns
This is primarily to reduce build times and length of the main stablehlo
to linalg conversion pass file.

Also clean up the code:
-  Use `LogicalResult` for verification success/failure
-  Use free cast functions.
-  Replace some local `auto` with the actual type where not obvious.

Issue: https://github.com/openxla/iree/issues/12678
