+++ compiler/src/iree/compiler/InputConversion/StableHLO/BUILD.bazel
@@ -49,6 +49,7 @@ iree_compiler_cc_library(
         "LegalizeToLinalgUtils.cpp",
         "Passes.cpp",
         "StableHLOToLinalg.cpp",
+        "StableHLOToLinalgDotProd.cpp",
         "StableHLOToLinalgPointwise.cpp",
         "TypeConversion.cpp",
     ],
